LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/img.https.sub-expected.txt

Blocked access to external URL https://www.localhost:9443/common/security-features/subresource/image.py


PASS Same-origin image
FAIL Same-site image promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Cross-site image promise_test: Unhandled rejection with value: object "[object Event]"
PASS Same-origin image: sec-fetch-dest
PASS Same-origin image: sec-fetch-mode
FAIL Same-origin image: sec-fetch-site assert_equals: expected (string) "same-origin" but got (undefined) undefined
PASS Same-origin image: sec-fetch-user
